Load edit form with pjax.
From http://pjax.heroku.com:

    pjax loads html from your server into the current page without
    a full page load. It's ajax with real permalinks, page titles,
    and a working back button that fully degrades

+++ app/assets/javascripts/controls.js.coffee
@@ -1,6 +1,15 @@
 $ ->
-  resource = location.pathname
-  $('.alert-message .close').click ->
-    id = $(@).parent('.alert-message').hide().attr('id')
-    $.ajax("#{resource}/alerts/#{id}", {type: 'delete'})
+  # close alerts
+  resource = location.pathname.match(/^\/profiles\/\d+/)[0]
+  $('.alert-message .close').live 'click', ->
+    id = $(@).parents('.alert-message').hide().attr('id')
+    $.ajax("#{resource}/alerts/#{id}", {type: 'delete'}) if id
     false
+
+  $('#aux .close, #aux .close-btn').live 'click', ->
+    $(@).parents('#aux').empty()
+    history.pushState({}, "profile", resource) if history.pushState
+    false
+
+  # pjax
+  $('a[data-pjax]').pjax()
